Hmmm..
If I have
   Float someFloat = ... ;
   if (someFloat == 1f) {

... is someFloat unboxed for comparison to 1f, or is 1f boxed and
the object references compared? I gotta imagine it does the right thing,
but let's be explicit. I'll test this out quickly.


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@767 a9e240f3-4568-de11-90a4-003048321218

+++ src/java/com/threerings/everything/server/ThingIndex.java
@@ -390,7 +390,9 @@ public class ThingIndex
         public ThingInfo copyWeighted (IntMap<Float> weights)
         {
             Float adjust = weights.get(categoryId);
-            return ((adjust == null) || (adjust == 1f)) ? this : new ThingInfo(this, adjust);
+            return ((adjust == null) || (adjust.floatValue() == 1f))
+                ? this
+                : new ThingInfo(this, adjust);
         }
 
         public String toString () {
